backend/requirements.txt
requests==2.31.0
pandas>=2.2.0
dash==2.17.1
plotly==5.22.0
dash-bootstrap-components==1.6.0
flask==3.0.0
flask-cors==4.0.0
gunicorn==22.0.0
python-dotenv==1.0.0
jupyter==1.0.0
ipykernel>=6.29.0